install: split mod.rs into bin_linking/git_prepare/lifecycle submodules
Pull three cohesive helper groups out of the 4847-line install/mod.rs
into their own files so the top-level pipeline is easier to navigate:

- bin_linking.rs: materialized_pkg_dir, link_bins, link_dep_bins,
  link_bundled_bins, link_bin_entries, and their helpers
- git_prepare.rs: ScratchDir, prepare_scratch_copy, run_git_dep_prepare
  and the nested-prepare depth cap
- lifecycle.rs: run_root_lifecycle, build_policy_from_sources,
  resolve_link_strategy, run_dep_lifecycle_scripts,
  import_verified_tarball, validate_required_scripts,
  unreviewed_dep_builds

mod.rs drops from 4847 to 3737 lines. No behavior changes; pub(crate)
functions stay accessible via re-exports so rebuild.rs,
ignored_builds.rs, and state.rs call sites are untouched.
