+++ crates/aube/src/commands/install/mod.rs
@@ -7,15 +7,26 @@ use miette::{Context, IntoDiagnostic, miette};
 use rayon::prelude::*;
 use std::collections::BTreeMap;
 
+mod bin_linking;
 mod delta;
 mod dep_selection;
 mod frozen;
+mod git_prepare;
+mod lifecycle;
 mod node_gyp_bootstrap;
 mod settings;
 mod side_effects_cache;
 
+use bin_linking::{link_bin_entries, link_bins, link_bins_for_dep};
+pub(crate) use bin_linking::{link_dep_bins, materialized_pkg_dir};
 pub use dep_selection::DepSelection;
 pub use frozen::{FrozenMode, FrozenOverride, GlobalVirtualStoreFlags};
+use git_prepare::{prepare_scratch_copy, run_git_dep_prepare};
+pub(crate) use lifecycle::{build_policy_from_sources, run_dep_lifecycle_scripts};
+use lifecycle::{
+    import_verified_tarball, resolve_link_strategy, run_root_lifecycle, unreviewed_dep_builds,
+    validate_required_scripts,
+};
 pub(crate) use settings::PeerDependencyRules;
 pub(crate) use settings::resolve_dependency_policy;
 pub(crate) use side_effects_cache::{SideEffectsCacheConfig, side_effects_cache_root};
@@ -31,7 +42,6 @@ use settings::{
     resolve_strict_store_pkg_content_check, resolve_symlink, resolve_use_running_store_server,
     resolve_verify_store_integrity,
 };
-use side_effects_cache::{SideEffectsCacheEntry, SideEffectsCacheRestore};
 
 #[derive(Debug, clap::Args)]
 pub struct InstallArgs {
@@ -396,694 +406,6 @@ impl From<FrozenMode> for InstallOptions {
     fn from(mode: FrozenMode) -> Self {
         Self::with_mode(mode)
     }
-}
-
-/// Run a root-package lifecycle hook, announcing it to the user if defined
-/// and turning aube_scripts::Error into a miette::Report with context.
-/// Silent when the hook isn't defined in package.json.
-async fn run_root_lifecycle(
-    project_dir: &std::path::Path,
-    modules_dir_name: &str,
-    manifest: &aube_manifest::PackageJson,
-    hook: aube_scripts::LifecycleHook,
-) -> miette::Result<()> {
-    // Only announce when the hook is actually defined, so projects without
-    // lifecycle scripts don't get noise in their install output.
-    if !manifest.scripts.contains_key(hook.script_name()) {
-        return Ok(());
-    }
-    tracing::debug!("Running {} script...", hook.script_name());
-    aube_scripts::run_root_hook(project_dir, modules_dir_name, manifest, hook)
-        .await
-        .map_err(|e| {
-            // Old message was just the bare error string. User got
-            // a cryptic "exit status 1" with no hook name, no script
-            // path, nothing. Tag with which hook fired so the log
-            // line is self-documenting. This is the common case
-            // (failed preinstall on `aube install`) so the regression
-            // really hurt triage.
-            miette!("root {} script failed: {e}", hook.script_name())
-        })?;
-    Ok(())
-}
-
-/// Build the dependency lifecycle-script `BuildPolicy` by merging
-/// every supported source on the root manifest + workspace file:
-///
-/// - `package.json` / `pnpm-workspace.yaml` `pnpm.allowBuilds` map
-///   (aube's superset format — patterns with bool values)
-/// - `package.json` / `pnpm-workspace.yaml` `pnpm.onlyBuiltDependencies`
-///   flat list (pnpm's canonical allowlist, used by nearly every
-///   real-world pnpm project)
-/// - `package.json` / `pnpm-workspace.yaml` `pnpm.neverBuiltDependencies`
-///   flat list (pnpm's canonical denylist)
-/// - the `--dangerously-allow-all-builds` escape hatch
-///
-/// Workspace-level entries in the `allowBuilds` map take precedence
-/// over the manifest map for the same pattern, matching pnpm. The
-/// flat lists are pure append — deny always wins at `decide()` time.
-pub(crate) fn build_policy_from_sources(
-    manifest: &aube_manifest::PackageJson,
-    workspace: &aube_manifest::WorkspaceConfig,
-    dangerously_allow_all_builds: bool,
-) -> (
-    aube_scripts::BuildPolicy,
-    Vec<aube_scripts::BuildPolicyError>,
-) {
-    let mut merged = manifest.pnpm_allow_builds();
-    for (k, v) in workspace.allow_builds_raw() {
-        merged.insert(k, v);
-    }
-    let mut only_built = manifest.pnpm_only_built_dependencies();
-    only_built.extend(workspace.only_built_dependencies.iter().cloned());
-    // Bun's top-level `trustedDependencies` feeds the same allowlist so
-    // bun projects migrating to aube keep running their install scripts
-    // without moving the list under `pnpm.onlyBuiltDependencies` first.
-    only_built.extend(manifest.trusted_dependencies());
-    let mut never_built = manifest.pnpm_never_built_dependencies();
-    never_built.extend(workspace.never_built_dependencies.iter().cloned());
-    aube_scripts::BuildPolicy::from_config(
-        &merged,
-        &only_built,
-        &never_built,
-        dangerously_allow_all_builds,
-    )
-}
-
-/// Resolve the link strategy (reflink / hardlink / copy) from CLI
-/// override, `.npmrc` / `pnpm-workspace.yaml`, or filesystem detection.
-/// Shared by the prewarm-GVS materializer (which needs the strategy
-/// before the full linker is built) and the link phase proper.
-pub(crate) fn resolve_link_strategy(
-    cwd: &std::path::Path,
-    ctx: &aube_settings::ResolveCtx<'_>,
-) -> miette::Result<aube_linker::LinkStrategy> {
-    let package_import_method_cli =
-        aube_settings::values::string_from_cli("packageImportMethod", ctx.cli);
-    // Shared probe used by both the CLI and resolved-setting paths
-    // below. Probe across store dir and project modules dir. Single
-    // dir probe reports reflink but real link ops across a mount
-    // boundary hit EXDEV and silently fall back to per-file copy.
-    // Catches the cross-FS case at probe time.
-    let auto_probe = || {
-        let store_dir = super::open_store(cwd).map(|s| s.root().to_path_buf()).ok();
-        let modules_dir = cwd.join("node_modules");
-        match store_dir.as_deref() {
-            Some(sd) => aube_linker::Linker::detect_strategy_cross(sd, &modules_dir),
-            None => aube_linker::Linker::detect_strategy(cwd),
-        }
-    };
-    let strategy = if let Some(cli) = package_import_method_cli.as_deref() {
-        match cli.trim().to_ascii_lowercase().as_str() {
-            "" | "auto" => auto_probe(),
-            "hardlink" => aube_linker::LinkStrategy::Hardlink,
-            "copy" => aube_linker::LinkStrategy::Copy,
-            "clone-or-copy" => aube_linker::LinkStrategy::Reflink,
-            "clone" => {
-                tracing::warn!(
-                    "package-import-method=clone: reflink will silently fall back to copy \
-                     if the filesystem does not support it (strict enforcement is a known TODO)"
-                );
-                aube_linker::LinkStrategy::Reflink
-            }
-            other => {
-                return Err(miette!(
-                    "unknown --package-import-method value `{other}`; expected `auto`, `hardlink`, `copy`, `clone`, or `clone-or-copy`"
-                ));
-            }
-        }
-    } else {
-        match aube_settings::resolved::package_import_method(ctx) {
-            aube_settings::resolved::PackageImportMethod::Auto => auto_probe(),
-            aube_settings::resolved::PackageImportMethod::Hardlink => {
-                aube_linker::LinkStrategy::Hardlink
-            }
-            aube_settings::resolved::PackageImportMethod::Copy => aube_linker::LinkStrategy::Copy,
-            aube_settings::resolved::PackageImportMethod::CloneOrCopy => {
-                aube_linker::LinkStrategy::Reflink
-            }
-            aube_settings::resolved::PackageImportMethod::Clone => {
-                tracing::warn!(
-                    "package-import-method=clone: reflink will silently fall back to copy \
-                     if the filesystem does not support it (strict enforcement is a known TODO)"
-                );
-                aube_linker::LinkStrategy::Reflink
-            }
-        }
-    };
-    Ok(strategy)
-}
-
-/// Walk every linked dependency, check its `package.json` for
-/// lifecycle scripts, and run the ones the policy allows. Runs
-/// `preinstall` → `install` → `postinstall` per package in that order;
-/// `prepare` is skipped for deps (pnpm does the same).
-///
-/// `package_indices` gives us the stored `package.json` for each dep
-/// without a second disk read, and the actual execution cwd is
-/// `node_modules/.aube/<dep_path>/node_modules/<name>` — i.e. the
-/// linked dir inside the virtual store.
-#[allow(clippy::too_many_arguments)]
-pub(crate) async fn run_dep_lifecycle_scripts(
-    project_dir: &std::path::Path,
-    modules_dir_name: &str,
-    aube_dir: &std::path::Path,
-    graph: &aube_lockfile::LockfileGraph,
-    policy: &aube_scripts::BuildPolicy,
-    virtual_store_dir_max_length: usize,
-    child_concurrency: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-    side_effects_cache: SideEffectsCacheConfig<'_>,
-) -> miette::Result<usize> {
-    // Pass 1 (serial, cheap): walk the graph, keep only the packages
-    // the policy allows AND that actually define at least one dep
-    // lifecycle hook in their on-disk `package.json`. Filtering up front
-    // means the fan-out below only spawns real work — no tokio task per
-    // every 200-package graph for a graph that has 3 allowlisted deps.
-    #[derive(Clone)]
-    struct BuildJob {
-        name: String,
-        version: String,
-        package_dir: std::path::PathBuf,
-        /// Directory containing the dep package and its sibling
-        /// symlinks — i.e. `package_dir`'s enclosing `node_modules/`.
-        /// `<dep_modules_dir>/.bin` is prepended to PATH so the
-        /// postinstall can call binaries declared in the dep's own
-        /// `dependencies`. See `link_dep_bins` for the write side.
-        dep_modules_dir: std::path::PathBuf,
-        manifest: aube_manifest::PackageJson,
-        cache_entry: Option<SideEffectsCacheEntry>,
-    }
-
-    let mut jobs: Vec<BuildJob> = Vec::new();
-    for (dep_path, pkg) in &graph.packages {
-        // Use registry_name(), not pkg.name. pkg.name is the in-tree
-        // alias (`h3-safe`). Real package is `h3`. Allowlist entry for
-        // `h3` would miss if we checked against the alias. Attacker
-        // writes `"h3-safe": "npm:h3@0.19.0"` to sneak a denied pkg
-        // through the allowlist. registry_name() strips alias back to
-        // real name.
-        match policy.decide(pkg.registry_name(), &pkg.version) {
-            aube_scripts::AllowDecision::Allow => {}
-            aube_scripts::AllowDecision::Deny | aube_scripts::AllowDecision::Unspecified => {
-                continue;
-            }
-        }
-        let package_dir = materialized_pkg_dir(
-            aube_dir,
-            dep_path,
-            &pkg.name,
-            virtual_store_dir_max_length,
-            placements,
-        );
-        if !package_dir.exists() {
-            tracing::debug!(
-                "allowBuilds: skipping {} — {} not on disk",
-                pkg.name,
-                package_dir.display()
-            );
-            continue;
-        }
-        // Read the dep's `package.json` directly from its materialized
-        // location. Previously we looked it up via `package_indices`,
-        // but the fetch phase now skips `load_index` for packages
-        // whose virtual-store entry already exists (which is every
-        // package on a no-op re-install), so the map is sparse and
-        // many dep_paths legitimately won't have an entry. The
-        // on-disk file is hardlinked to the same bytes the store
-        // would have pointed us at.
-        //
-        // `NotFound` is the only error we swallow here: some packages
-        // legitimately ship without a top-level `package.json` (or
-        // the field gets stripped by linkers that treat the virtual
-        // store as opaque), and we shouldn't fail the install over
-        // that. Every other I/O error — permission denied, disk
-        // corruption, short reads — surfaces as a hard failure so
-        // the user sees the real problem instead of a silently
-        // skipped `node-gyp rebuild` or similar.
-        let pkg_json_path = package_dir.join("package.json");
-        let pkg_json_content = match std::fs::read_to_string(&pkg_json_path) {
-            Ok(s) => s,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => continue,
-            Err(e) => {
-                return Err(miette!(
-                    "failed to read package.json for {} at {}: {}",
-                    pkg.name,
-                    pkg_json_path.display(),
-                    e
-                ));
-            }
-        };
-        let dep_manifest = aube_manifest::PackageJson::parse(&pkg_json_path, pkg_json_content)
-            .map_err(miette::Report::new)
-            .wrap_err_with(|| format!("failed to parse package.json for {}", pkg.name))?;
-        // `has_dep_lifecycle_work` also accounts for the implicit
-        // `node-gyp rebuild` fallback: a package with a top-level
-        // `binding.gyp` and no `install`/`preinstall` script still has
-        // work to run, and pre-filtering on `scripts` alone would drop
-        // it before the fan-out even saw it.
-        if !aube_scripts::has_dep_lifecycle_work(&package_dir, &dep_manifest) {
-            continue;
-        }
-        let cache_entry = side_effects_cache
-            .root()
-            .map(|root| SideEffectsCacheEntry::new(root, &pkg.name, &pkg.version, &package_dir))
-            .transpose()?;
-        let dep_modules_dir = dep_modules_dir_for(&package_dir, &pkg.name);
-        jobs.push(BuildJob {
-            name: pkg.name.clone(),
-            version: pkg.version.clone(),
-            package_dir,
-            dep_modules_dir,
-            manifest: dep_manifest,
-            cache_entry,
-        });
-    }
-
-    if jobs.is_empty() {
-        return Ok(0);
-    }
-
-    // Bootstrap node-gyp once before the fan-out when the ambient
-    // `PATH` doesn't already provide one. At least one job is about
-    // to run a lifecycle script, and we can't cheaply predict which
-    // ones will end up shelling out to `node-gyp` (explicit,
-    // implicit via binding.gyp, or transitive via node-gyp-build).
-    // If the user already has node-gyp (system install, nvm, a test
-    // shim), `ensure` returns `None` and we leave their copy alone.
-    let node_gyp_bin_dir = std::sync::Arc::new(node_gyp_bootstrap::ensure(project_dir).await?);
-
-    // Pass 2 (parallel, bounded): fan out across `child_concurrency`
-    // concurrent workers. Inside one job the three hooks
-    // (preinstall → install → postinstall) still run sequentially —
-    // pnpm's execution model is "at most N packages building in
-    // parallel," not "at most N scripts running," so hook ordering
-    // within a single package is preserved.
-    //
-    // Cancellation on first failure uses `JoinSet`, which aborts every
-    // outstanding task when it's dropped. A plain `Vec<JoinHandle>`
-    // would NOT be safe here — dropping a `tokio::spawn` handle lets
-    // the task keep running detached, so a failing script would
-    // silently leave N siblings still executing `postinstall` against
-    // the user's machine after the install returned an error.
-    // `join_next` also surfaces whichever task fails first rather than
-    // waiting for the longest-running one to finish.
-    let concurrency = child_concurrency.max(1);
-    let semaphore = std::sync::Arc::new(tokio::sync::Semaphore::new(concurrency));
-    let project_dir = project_dir.to_path_buf();
-    let modules_dir_name = modules_dir_name.to_string();
-    let should_restore_side_effects_cache = side_effects_cache.should_restore();
-    let should_save_side_effects_cache = side_effects_cache.should_save();
-    let overwrite_side_effects_cache = side_effects_cache.overwrite_existing();
-    let mut set: tokio::task::JoinSet<miette::Result<usize>> = tokio::task::JoinSet::new();
-    for job in jobs {
-        let sem = semaphore.clone();
-        let project_dir = project_dir.clone();
-        let modules_dir_name = modules_dir_name.clone();
-        let node_gyp_bin_dir = node_gyp_bin_dir.clone();
-        set.spawn(async move {
-            let _permit = sem.acquire().await.unwrap();
-            if should_restore_side_effects_cache && let Some(cache_entry) = job.cache_entry.clone()
-            {
-                let package_dir = job.package_dir.clone();
-                let restore_result = tokio::task::spawn_blocking(move || {
-                    cache_entry.restore_if_available(&package_dir)
-                })
-                .await
-                .map_err(|e| {
-                    miette!(
-                        "side-effects-cache restore task panicked for {}@{}: {e}",
-                        job.name,
-                        job.version
-                    )
-                })?;
-                match restore_result? {
-                    SideEffectsCacheRestore::Restored | SideEffectsCacheRestore::AlreadyApplied => {
-                        return Ok(0);
-                    }
-                    SideEffectsCacheRestore::Miss => {}
-                }
-            }
-            let tool_dirs: Vec<&std::path::Path> = node_gyp_bin_dir
-                .as_ref()
-                .as_deref()
-                .map(|p| vec![p])
-                .unwrap_or_default();
-            let mut ran_here = 0usize;
-            for hook in aube_scripts::DEP_LIFECYCLE_HOOKS {
-                let did_run = aube_scripts::run_dep_hook(
-                    &job.package_dir,
-                    &job.dep_modules_dir,
-                    &project_dir,
-                    &modules_dir_name,
-                    &job.manifest,
-                    hook,
-                    &tool_dirs,
-                )
-                .await
-                .map_err(|e| {
-                    miette!(
-                        "lifecycle script {} failed for {}@{}: {}",
-                        hook.script_name(),
-                        job.name,
-                        job.version,
-                        e
-                    )
-                })?;
-                if did_run {
-                    tracing::debug!(
-                        "ran {} for {}@{}",
-                        hook.script_name(),
-                        job.name,
-                        job.version
-                    );
-                    ran_here += 1;
-                }
-            }
-            if should_save_side_effects_cache
-                && ran_here > 0
-                && let Some(cache_entry) = job.cache_entry.clone()
-            {
-                let package_dir = job.package_dir.clone();
-                let save_result = tokio::task::spawn_blocking(move || {
-                    cache_entry.save(&package_dir, overwrite_side_effects_cache)
-                })
-                .await
-                .map_err(|e| {
-                    miette!(
-                        "side-effects-cache save task panicked for {}@{}: {e}",
-                        job.name,
-                        job.version
-                    )
-                })
-                .and_then(|r| r);
-                if let Err(e) = save_result {
-                    tracing::debug!(
-                        "side-effects-cache: ignoring cache save error for {}@{}: {e}",
-                        job.name,
-                        job.version
-                    );
-                }
-            }
-            Ok(ran_here)
-        });
-    }
-
-    let mut ran = 0usize;
-    while let Some(res) = set.join_next().await {
-        // `?` on the outer `Result` propagates a real task-level panic
-        // (tokio's `JoinError`); `?` on the inner `miette::Result`
-        // propagates a script failure. Either way, the function
-        // returns, `set` is dropped, and the remaining in-flight
-        // scripts are aborted before they can scribble on disk.
-        ran += res.into_diagnostic()??;
-    }
-    Ok(ran)
-}
-
-/// Verify + import + validate + save-index for a freshly fetched
-/// tarball. Shared between the lockfile-driven fetch path and the
-/// no-lockfile streaming fetch path so both honor the same integrity
-/// and content-check settings. Runs inside `spawn_blocking` — no
-/// async in this function.
-#[allow(clippy::too_many_arguments)]
-fn import_verified_tarball(
-    store: &aube_store::Store,
-    bytes: &[u8],
-    display_name: &str,
-    registry_name: &str,
-    version: &str,
-    integrity: Option<&str>,
-    verify_integrity: bool,
-    strict_integrity: bool,
-    strict_pkg_content_check: bool,
-) -> miette::Result<aube_store::PackageIndex> {
-    if verify_integrity {
-        if let Some(expected) = integrity {
-            aube_store::verify_integrity(bytes, expected)
-                .map_err(|e| miette!("{display_name}@{version}: {e}"))?;
-        } else if strict_integrity {
-            // strict-store-integrity=true opts the user into
-            // fail-closed. Default is off so ecosystem parity with
-            // pnpm stays intact. A registry proxy that strips
-            // dist.integrity will no longer slip past silently when
-            // strict is on.
-            return Err(miette!(
-                "{display_name}@{version}: registry response has no `dist.integrity` and `strict-store-integrity` is on. Refusing to import unverified bytes."
-            ));
-        } else {
-            tracing::warn!(
-                "{display_name}@{version}: registry response has no `dist.integrity`, importing without content verification. Set `strict-store-integrity=true` to refuse instead."
-            );
-        }
-    }
-    let index = store
-        .import_tarball(bytes)
-        .map_err(|e| miette!("failed to import {display_name}@{version}: {e}"))?;
-    // strictStorePkgContentCheck: cross-check the freshly stored
-    // package.json against the resolver-asserted (name, version)
-    // before the index is cached or returned to the linker. Validate
-    // against `registry_name` — the real package name that appears
-    // in the tarball's own `package.json` — not the alias, or this
-    // would fail every npm-aliased entry.
-    if strict_pkg_content_check {
-        aube_store::validate_pkg_content(&index, registry_name, version)
-            .map_err(|e| miette!("{display_name}@{version}: {e}"))?;
-    }
-    // Cache under `registry_name` so two aliases of the same real
-    // package hit the same on-disk index file and avoid redundant
-    // fetches. When `integrity` is `Some` the filename carries a
-    // `+<hex>` suffix that discriminates same-(name, version)
-    // tarballs from different sources; when `None` falls back to the
-    // plain name@version key so warm installs still find the cache
-    // on integrity-stripping proxies.
-    if let Err(e) = store.save_index(registry_name, version, integrity, &index) {
-        tracing::warn!("Failed to cache index for {display_name}@{version}: {e}");
-    }
-    Ok(index)
-}
-
-fn validate_required_scripts(
-    project_dir: &std::path::Path,
-    manifest: &aube_manifest::PackageJson,
-    required: &[String],
-) -> miette::Result<()> {
-    if required.is_empty() {
-        return Ok(());
-    }
-    let mut missing = Vec::new();
-    collect_missing_required_scripts(".", manifest, required, &mut missing);
-    for pkg_dir in aube_workspace::find_workspace_packages(project_dir)
-        .map_err(|e| miette!("failed to discover workspace packages: {e}"))?
-    {
-        let manifest_path = pkg_dir.join("package.json");
-        let pkg_manifest = aube_manifest::PackageJson::from_path(&manifest_path)
-            .map_err(miette::Report::new)
-            .wrap_err_with(|| format!("failed to read {}", manifest_path.display()))?;
-        let label = pkg_manifest
-            .name
-            .as_deref()
-            .map(str::to_string)
-            .unwrap_or_else(|| {
-                pkg_dir
-                    .strip_prefix(project_dir)
-                    .unwrap_or(&pkg_dir)
-                    .display()
-                    .to_string()
-            });
-        collect_missing_required_scripts(&label, &pkg_manifest, required, &mut missing);
-    }
-    if missing.is_empty() {
-        Ok(())
-    } else {
-        Err(miette!(
-            "requiredScripts check failed:\n{}",
-            missing
-                .into_iter()
-                .map(|(pkg, script)| format!("  - {pkg} is missing `{script}`"))
-                .collect::<Vec<_>>()
-                .join("\n")
-        ))
-    }
-}
-
-fn collect_missing_required_scripts(
-    label: &str,
-    manifest: &aube_manifest::PackageJson,
-    required: &[String],
-    missing: &mut Vec<(String, String)>,
-) {
-    for script in required {
-        if !manifest.scripts.contains_key(script) {
-            missing.push((label.to_string(), script.clone()));
-        }
-    }
-}
-
-fn unreviewed_dep_builds(
-    aube_dir: &std::path::Path,
-    graph: &aube_lockfile::LockfileGraph,
-    policy: &aube_scripts::BuildPolicy,
-    virtual_store_dir_max_length: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-) -> miette::Result<Vec<String>> {
-    let mut unreviewed = Vec::new();
-    for (dep_path, pkg) in &graph.packages {
-        if !matches!(
-            policy.decide(pkg.registry_name(), &pkg.version),
-            aube_scripts::AllowDecision::Unspecified
-        ) {
-            continue;
-        }
-        let package_dir = materialized_pkg_dir(
-            aube_dir,
-            dep_path,
-            &pkg.name,
-            virtual_store_dir_max_length,
-            placements,
-        );
-        if !package_dir.exists() {
-            continue;
-        }
-        let pkg_json_path = package_dir.join("package.json");
-        let pkg_json_content = match std::fs::read_to_string(&pkg_json_path) {
-            Ok(s) => s,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => continue,
-            Err(e) => {
-                return Err(miette!(
-                    "failed to read package.json for {} at {}: {}",
-                    pkg.name,
-                    pkg_json_path.display(),
-                    e
-                ));
-            }
-        };
-        let dep_manifest = aube_manifest::PackageJson::parse(&pkg_json_path, pkg_json_content)
-            .map_err(miette::Report::new)
-            .wrap_err_with(|| format!("failed to parse package.json for {}", pkg.name))?;
-        if aube_scripts::has_dep_lifecycle_work(&package_dir, &dep_manifest) {
-            unreviewed.push(pkg.spec_key());
-        }
-    }
-    unreviewed.sort();
-    unreviewed.dedup();
-    Ok(unreviewed)
-}
-
-/// Unique-per-call scratch directory that `rm -rf`s itself on drop.
-/// Used to run a git dep's `prepare` script without mutating the
-/// shared `git_shallow_clone` cache under `/tmp/aube-git-*`.
-struct ScratchDir(std::path::PathBuf);
-
-impl ScratchDir {
-    fn path(&self) -> &std::path::Path {
-        &self.0
-    }
-}
-
-impl Drop for ScratchDir {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_dir_all(&self.0);
-    }
-}
-
-/// Recursively copy `src` into a fresh temp directory and return it
-/// wrapped in a [`ScratchDir`]. `.git/` is intentionally skipped —
-/// prepare scripts never need the history, and dropping it keeps the
-/// copy an order of magnitude smaller on large repos. Uses `cp -a`
-/// so symlinks + file modes survive (matters for repos that ship
-/// executable bits their prepare script relies on).
-fn prepare_scratch_copy(src: &std::path::Path, spec: &str) -> miette::Result<ScratchDir> {
-    let mut hasher = std::collections::hash_map::DefaultHasher::new();
-    use std::hash::{Hash, Hasher};
-    src.hash(&mut hasher);
-    std::process::id().hash(&mut hasher);
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0)
-        .hash(&mut hasher);
-    let dst = std::env::temp_dir().join(format!("aube-git-prep-{:x}", hasher.finish()));
-    if dst.exists() {
-        let _ = std::fs::remove_dir_all(&dst);
-    }
-    std::fs::create_dir_all(&dst)
-        .map_err(|e| miette!("git dep {spec}: create scratch dir {}: {e}", dst.display()))?;
-
-    // Wrap the directory in `ScratchDir` *before* running any of
-    // the fallible work below. Handing ownership of cleanup to
-    // the Drop impl immediately means a failure to spawn `cp`, a
-    // non-zero cp exit, or any panic between here and the `Ok`
-    // return still removes the partially-populated temp dir
-    // instead of leaking it under `/tmp/aube-git-prep-*`.
-    let scratch = ScratchDir(dst);
-
-    // `cp -a src/. dst/` — the trailing `/.` copies src's contents
-    // (including dotfiles) into dst rather than creating `dst/<src>`.
-    // `-a` preserves perms/symlinks/timestamps. We exclude `.git`
-    // manually afterwards rather than with `--exclude` (non-POSIX,
-    // GNU-only).
-    let out = std::process::Command::new("cp")
-        .arg("-a")
-        .arg(format!("{}/.", src.display()))
-        .arg(scratch.path())
-        .output()
-        .map_err(|e| miette!("git dep {spec}: spawn cp for scratch copy: {e}"))?;
-    if !out.status.success() {
-        return Err(miette!(
-            "git dep {spec}: scratch copy failed: {}",
-            String::from_utf8_lossy(&out.stderr).trim()
-        ));
-    }
-    let _ = std::fs::remove_dir_all(scratch.path().join(".git"));
-
-    Ok(scratch)
-}
-
-/// Hard cap for nested git dep `prepare` installs. Four levels is more
-/// than any real-world chain we've seen and prevents a pathological repo
-/// from wedging install in an infinite clone loop.
-const GIT_PREPARE_MAX_DEPTH: u32 = 4;
-
-/// Run a nested `aube install` inside a git-dep checkout so its
-/// devDependencies are linked and its root `prepare` script runs
-/// before the caller snapshots the tree via `aube pack`.
-///
-/// `ignore_scripts` is forwarded from the outer install so a user
-/// who passed `--ignore-scripts` for security/reproducibility
-/// reasons doesn't have the git dep's full root lifecycle sequence
-/// execute regardless — the caller is expected to *skip* calling
-/// this function entirely under `--ignore-scripts`, but we still
-/// forward the flag as a belt-and-suspenders defense in case a
-/// nested install reaches this path through some other code path.
-async fn run_git_dep_prepare(
-    clone_dir: &std::path::Path,
-    spec: &str,
-    ignore_scripts: bool,
-    depth: u32,
-) -> miette::Result<()> {
-    if depth >= GIT_PREPARE_MAX_DEPTH {
-        return Err(miette!(
-            "git dep {spec}: `prepare` nesting exceeded {GIT_PREPARE_MAX_DEPTH} levels"
-        ));
-    }
-    let mut opts = InstallOptions::with_mode(super::chained_frozen_mode(FrozenMode::Prefer));
-    opts.project_dir = Some(clone_dir.to_path_buf());
-    opts.ignore_scripts = ignore_scripts;
-    opts.git_prepare_depth = depth + 1;
-    let spec = spec.to_string();
-    tokio::task::spawn_blocking(move || {
-        let runtime = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .into_diagnostic()
-            .wrap_err("failed to build nested git prepare runtime")?;
-        runtime.block_on(run(opts))
-    })
-    .await
-    .into_diagnostic()
-    .wrap_err_with(|| format!("git dep {spec}: nested install task failed"))?
-    .wrap_err_with(|| format!("git dep {spec}: nested install for `prepare` failed"))
 }
 
 /// Materialize a `file:` / `link:` package into the store.
@@ -4412,436 +3734,4 @@ fn filter_graph_to_importers<const N: usize>(
         ..graph.clone()
     };
     filtered.filter_deps(|_| true)
-}
-
-/// Link bin entries from packages to node_modules/.bin/
-/// Compute the on-disk directory a dep's materialized package lives
-/// in. Matches the path `aube-linker` writes under
-/// `node_modules/.aube/<escaped dep_path>/node_modules/<name>`.
-///
-/// `virtual_store_dir_max_length` must match the value the linker
-/// was built with (see `install::run` for the single source of
-/// truth) — otherwise long `dep_path`s that trigger the
-/// truncate-and-hash fallback inside `dep_path_to_filename` will
-/// encode to a different filename than the one the linker wrote,
-/// and this function will return a path that doesn't exist.
-pub(crate) fn materialized_pkg_dir(
-    aube_dir: &std::path::Path,
-    dep_path: &str,
-    name: &str,
-    virtual_store_dir_max_length: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-) -> std::path::PathBuf {
-    // In hoisted mode the package was materialized directly into
-    // `node_modules/<...>/<name>/` and its path is recorded in
-    // `placements`. Fall back to the isolated `.aube/<dep_path>`
-    // convention when either the mode is isolated (`placements` is
-    // `None`) or the hoisted planner didn't place this specific
-    // dep_path (e.g. filtered by `--prod` / `--no-optional`).
-    // `aube_dir` is the resolved `virtualStoreDir` — the install
-    // driver threads it in via `commands::resolve_virtual_store_dir`
-    // so a custom override lands on the same path the linker wrote
-    // to.
-    if let Some(placements) = placements
-        && let Some(p) = placements.package_dir(dep_path)
-    {
-        return p.to_path_buf();
-    }
-    aube_dir
-        .join(dep_path_to_filename(dep_path, virtual_store_dir_max_length))
-        .join("node_modules")
-        .join(name)
-}
-
-/// Directory holding the dep's own `node_modules/` — i.e. the dir
-/// that contains both `<name>` and its sibling symlinks. For scoped
-/// packages (`@scope/name`) `package_dir` is two levels below that
-/// `node_modules/`, so we strip the extra `@scope` hop. Used to
-/// locate the per-dep `.bin/` for transitive lifecycle-script bins.
-fn dep_modules_dir_for(package_dir: &std::path::Path, name: &str) -> std::path::PathBuf {
-    if name.starts_with('@') {
-        package_dir
-            .parent()
-            .and_then(std::path::Path::parent)
-            .map(std::path::Path::to_path_buf)
-            .unwrap_or_else(|| package_dir.to_path_buf())
-    } else {
-        package_dir
-            .parent()
-            .map(std::path::Path::to_path_buf)
-            .unwrap_or_else(|| package_dir.to_path_buf())
-    }
-}
-
-/// Read a dep's `package.json` from its materialized directory.
-///
-/// Earlier revisions of this file went through
-/// `package_indices[dep_path]` and read
-/// `stored.store_path.join("package.json")` from the CAS. That
-/// stopped working once `fetch_packages_with_root` learned to skip
-/// `load_index` for packages whose `.aube/<dep_path>` already exists
-/// (the `AlreadyLinked` fast path) — the indices map is sparse on
-/// warm installs, and every caller that reached for
-/// `package_indices.get(..)?.get("package.json")` silently dropped
-/// those deps via the `continue` or `?` on the missing key.
-///
-/// Read the hardlinked file at the materialized location instead:
-/// same bytes, zero dependency on the sparse indices map, and
-/// doesn't require a cache miss to surface when the virtual store is
-/// intact.
-///
-/// Error policy: `Ok(None)` only when the file is legitimately
-/// missing (e.g. a package that ships without a top-level
-/// `package.json`, or hasn't been materialized yet). Every other
-/// `std::io::Error` — permission denied, short reads, disk errors —
-/// bubbles up as `Err` so the user sees a real failure instead of a
-/// silently dropped bin link. Parse errors likewise propagate.
-fn read_materialized_pkg_json(
-    aube_dir: &std::path::Path,
-    dep_path: &str,
-    name: &str,
-    virtual_store_dir_max_length: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-) -> miette::Result<Option<serde_json::Value>> {
-    let pkg_dir = materialized_pkg_dir(
-        aube_dir,
-        dep_path,
-        name,
-        virtual_store_dir_max_length,
-        placements,
-    );
-    let pkg_json_path = pkg_dir.join("package.json");
-    let content = match std::fs::read_to_string(&pkg_json_path) {
-        Ok(s) => s,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-        Err(e) => {
-            return Err(miette!(
-                "failed to read package.json for {name} at {}: {e}",
-                pkg_json_path.display()
-            ));
-        }
-    };
-    let value = aube_manifest::parse_json::<serde_json::Value>(&pkg_json_path, content)
-        .map_err(miette::Report::new)
-        .wrap_err_with(|| format!("failed to parse package.json for {name}"))?;
-    Ok(Some(value))
-}
-
-/// Create top-level + bundled bin symlinks for one dep. Extracted so
-/// both the root-importer pass (`link_bins`) and the per-workspace
-/// loop use the same code path.
-#[allow(clippy::too_many_arguments)]
-fn link_bins_for_dep(
-    aube_dir: &std::path::Path,
-    bin_dir: &std::path::Path,
-    graph: &aube_lockfile::LockfileGraph,
-    dep_path: &str,
-    name: &str,
-    virtual_store_dir_max_length: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-    shim_opts: aube_linker::BinShimOptions,
-    has_bin_metadata: bool,
-) -> miette::Result<()> {
-    let pkg_dir = materialized_pkg_dir(
-        aube_dir,
-        dep_path,
-        name,
-        virtual_store_dir_max_length,
-        placements,
-    );
-    // Fast path: when the lockfile carries bin metadata and says this
-    // package ships none, skip the package.json read + JSON parse.
-    // 95%+ of a typical graph falls into this bucket; the saving
-    // scales with every bin-linking caller (root, per-dep,
-    // per-workspace). `local_source` packages (file:/link:) bypass
-    // the lockfile's bin info so we still consult their on-disk
-    // manifest. Bundled dependencies contribute bins from child
-    // tarballs regardless of the parent's own `bin` field, so
-    // `link_bundled_bins` runs unconditionally below.
-    let skip_bin_read = has_bin_metadata
-        && graph
-            .get_package(dep_path)
-            .is_some_and(|p| p.bin.is_empty() && p.local_source.is_none());
-    if skip_bin_read {
-        return link_bundled_bins(bin_dir, &pkg_dir, graph, dep_path, shim_opts);
-    }
-    if let Some(pkg_json) = read_materialized_pkg_json(
-        aube_dir,
-        dep_path,
-        name,
-        virtual_store_dir_max_length,
-        placements,
-    )? && let Some(bin) = pkg_json.get("bin")
-    {
-        link_bin_entries(bin_dir, &pkg_dir, Some(name), bin, shim_opts)?;
-    }
-    link_bundled_bins(bin_dir, &pkg_dir, graph, dep_path, shim_opts)?;
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-fn link_bins(
-    project_dir: &std::path::Path,
-    modules_dir_name: &str,
-    aube_dir: &std::path::Path,
-    graph: &aube_lockfile::LockfileGraph,
-    virtual_store_dir_max_length: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-    shim_opts: aube_linker::BinShimOptions,
-    has_bin_metadata: bool,
-) -> miette::Result<()> {
-    let bin_dir = project_dir.join(modules_dir_name).join(".bin");
-    std::fs::create_dir_all(&bin_dir).into_diagnostic()?;
-
-    for dep in graph.root_deps() {
-        link_bins_for_dep(
-            aube_dir,
-            &bin_dir,
-            graph,
-            &dep.dep_path,
-            &dep.name,
-            virtual_store_dir_max_length,
-            placements,
-            shim_opts,
-            has_bin_metadata,
-        )?;
-    }
-
-    Ok(())
-}
-
-/// Write per-dep `.bin/` directories holding shims for each package's
-/// *own* declared dependencies. Mirrors pnpm's post-link pass that
-/// populates `node_modules/.pnpm/<dep_path>/node_modules/.bin/`.
-///
-/// Without this, a dep's lifecycle script (e.g. `unrs-resolver`'s
-/// postinstall that calls `prebuild-install`) can't find transitive
-/// binaries on PATH — the project-level `node_modules/.bin` only holds
-/// shims for the root's *direct* deps. `run_dep_hook` prepends the
-/// dep-local `.bin` (via `dep_modules_dir_for`) before the
-/// project-level one so the dep's own transitive bins always win.
-///
-/// Isolated mode only. Hoisted mode materializes deps at the project
-/// root's `node_modules/` and generally relies on the single top-level
-/// `.bin`; nested transitive bins under hoisted are a known rough edge
-/// and out of scope here.
-pub(crate) fn link_dep_bins(
-    aube_dir: &std::path::Path,
-    graph: &aube_lockfile::LockfileGraph,
-    virtual_store_dir_max_length: usize,
-    placements: Option<&aube_linker::HoistedPlacements>,
-    shim_opts: aube_linker::BinShimOptions,
-    has_bin_metadata: bool,
-) -> miette::Result<()> {
-    if placements.is_some() {
-        // Hoisted — skip. See function doc.
-        return Ok(());
-    }
-    for (dep_path, pkg) in &graph.packages {
-        if pkg.dependencies.is_empty() {
-            continue;
-        }
-        // Fast path: when lockfile bin metadata is trustworthy and
-        // none of this package's children declare bins, the whole dep
-        // contributes nothing to a local `.bin/`. Skipping here avoids
-        // both the `pkg_dir.exists()` stat and the per-child
-        // `link_bins_for_dep` dispatch on ~95% of entries in a typical
-        // graph. Escape hatches: a child with `local_source`
-        // (file:/link:) bypasses the lockfile's bin map; a child with
-        // its *own* `bundled_dependencies` can ship bins from its
-        // nested tarballs that `link_bins_for_dep` -> `link_bundled_bins`
-        // surfaces into the parent's `.bin/`, so we must dispatch
-        // normally for it.
-        if has_bin_metadata
-            && pkg.bundled_dependencies.is_empty()
-            && pkg.dependencies.iter().all(|(child_name, child_version)| {
-                let child_dep_path = format!("{child_name}@{child_version}");
-                graph.get_package(&child_dep_path).is_some_and(|c| {
-                    c.bin.is_empty()
-                        && c.local_source.is_none()
-                        && c.bundled_dependencies.is_empty()
-                })
-            })
-        {
-            continue;
-        }
-        let pkg_dir = materialized_pkg_dir(
-            aube_dir,
-            dep_path,
-            &pkg.name,
-            virtual_store_dir_max_length,
-            placements,
-        );
-        if !pkg_dir.exists() {
-            // Filtered by optional / platform guards, or a staging
-            // hiccup. Skipping avoids blowing up the whole install on
-            // a dep that was never materialized.
-            continue;
-        }
-        let dep_modules_dir = dep_modules_dir_for(&pkg_dir, &pkg.name);
-        let bin_dir = dep_modules_dir.join(".bin");
-        // Don't `create_dir_all(&bin_dir)` here — most deps have
-        // no child that ships a `bin`, and an eager mkdir would leave
-        // empty `.bin/` directories everywhere. `create_bin_link`
-        // materializes the parent the first time a shim actually
-        // lands, so deps whose children contribute zero shims stay
-        // empty on disk.
-
-        for (child_name, child_version) in &pkg.dependencies {
-            // Mirror the linker's self-ref guard from
-            // `materialize_into`: a package that depends on its own
-            // dep_path is a graph artefact, not a real edge.
-            let child_dep_path = format!("{child_name}@{child_version}");
-            if child_dep_path == *dep_path && child_name == &pkg.name {
-                continue;
-            }
-            // The sibling may have been filtered (optional on another
-            // platform); `link_bins_for_dep` already returns Ok when
-            // the target pkg_json is absent, so just call through.
-            link_bins_for_dep(
-                aube_dir,
-                &bin_dir,
-                graph,
-                &child_dep_path,
-                child_name,
-                virtual_store_dir_max_length,
-                placements,
-                shim_opts,
-                has_bin_metadata,
-            )?;
-        }
-    }
-    Ok(())
-}
-
-/// Hoist bins declared by a package's `bundledDependencies` into
-/// `bin_dir`. The bundled children live under
-/// `<pkg_dir>/node_modules/<bundled>/` straight from the tarball — the
-/// resolver never walks them, so they don't show up in the regular
-/// packument-driven bin-linking pass and need this companion hoist.
-/// Matches pnpm's post-bin-linking pass for `hasBundledDependencies`.
-/// Used by both the root importer (`link_bins`) and the per-workspace
-/// loop so a workspace package depending on a parent with bundled deps
-/// sees the children's bins in its own `node_modules/.bin`.
-fn link_bundled_bins(
-    bin_dir: &std::path::Path,
-    pkg_dir: &std::path::Path,
-    graph: &aube_lockfile::LockfileGraph,
-    dep_path: &str,
-    shim_opts: aube_linker::BinShimOptions,
-) -> miette::Result<()> {
-    let Some(locked) = graph.get_package(dep_path) else {
-        return Ok(());
-    };
-    for bundled in &locked.bundled_dependencies {
-        let bundled_dir = pkg_dir.join("node_modules").join(bundled);
-        let bundled_pkg_json_path = bundled_dir.join("package.json");
-        let Ok(content) = std::fs::read_to_string(&bundled_pkg_json_path) else {
-            continue;
-        };
-        let Ok(bundled_pkg_json) = serde_json::from_str::<serde_json::Value>(&content) else {
-            continue;
-        };
-        let Some(bin) = bundled_pkg_json.get("bin") else {
-            continue;
-        };
-        link_bin_entries(bin_dir, &bundled_dir, Some(bundled), bin, shim_opts)?;
-    }
-    Ok(())
-}
-
-/// Shim each entry of a package.json `bin` field into `bin_dir`,
-/// resolving relative targets against `pkg_dir`. Shared by the
-/// dep-bin pass (`link_bins_for_dep`), bundled-deps pass
-/// (`link_bundled_bins`), and importer self-bin pass (root + each
-/// workspace member, discussion #228).
-///
-/// String-form `bin: "./x.js"` uses the basename of `pkg_name` as the
-/// shim name (scope `@a/b` → `b`); the entry is silently skipped when
-/// `pkg_name` is `None`. Object-form `bin: { foo: "./f" }` uses each
-/// key as-is. Entries whose name or target fail
-/// [`aube_linker::validate_bin_name`] / [`aube_linker::validate_bin_target`]
-/// are dropped without error, matching the pnpm/npm "silently ignore
-/// invalid bin" behavior.
-fn link_bin_entries(
-    bin_dir: &std::path::Path,
-    pkg_dir: &std::path::Path,
-    pkg_name: Option<&str>,
-    bin: &serde_json::Value,
-    shim_opts: aube_linker::BinShimOptions,
-) -> miette::Result<()> {
-    match bin {
-        serde_json::Value::String(bin_path) => {
-            let Some(name) = pkg_name else {
-                return Ok(());
-            };
-            let bin_name = name.split('/').next_back().unwrap_or(name);
-            if aube_linker::validate_bin_name(bin_name).is_ok()
-                && aube_linker::validate_bin_target(bin_path).is_ok()
-            {
-                create_bin_link(bin_dir, bin_name, &pkg_dir.join(bin_path), shim_opts)?;
-            }
-        }
-        serde_json::Value::Object(bins) => {
-            for (bin_name, path) in bins {
-                if let Some(path_str) = path.as_str()
-                    && aube_linker::validate_bin_name(bin_name).is_ok()
-                    && aube_linker::validate_bin_target(path_str).is_ok()
-                {
-                    create_bin_link(bin_dir, bin_name, &pkg_dir.join(path_str), shim_opts)?;
-                }
-            }
-        }
-        _ => {}
-    }
-    Ok(())
-}
-
-fn create_bin_link(
-    bin_dir: &std::path::Path,
-    name: &str,
-    target: &std::path::Path,
-    shim_opts: aube_linker::BinShimOptions,
-) -> miette::Result<()> {
-    // `link_dep_bins` skips eager `create_dir_all` on per-dep `.bin/`.
-    // Deps whose children ship no bins stay empty on disk. First shim
-    // write materializes the dir on demand.
-    //
-    // Windows `CreateDirectoryW` returns `ERROR_ALREADY_EXISTS` (os 183)
-    // when the leaf sits behind a junction in the path, even when the
-    // leaf is absent. The isolated layout's `.aube/<dep_path>` is a
-    // junction into the global virtual store, so every `.bin/` under it
-    // hits the quirk. Fix: canonicalize the parent, strip the `\\?\`
-    // UNC prefix (which trips its own CreateDirectoryW quirk, os 123),
-    // create the leaf on the plain drive path. No-op on Unix.
-    #[cfg(windows)]
-    let target_for_mkdir_owned = bin_dir.parent().and_then(|parent| {
-        let leaf = bin_dir.file_name()?;
-        let canon = std::fs::canonicalize(parent).ok()?;
-        let s = canon.to_string_lossy();
-        let cleaned = s.strip_prefix(r"\\?\").unwrap_or(&s);
-        Some(std::path::PathBuf::from(cleaned).join(leaf))
-    });
-    #[cfg(windows)]
-    let target_for_mkdir: &std::path::Path = target_for_mkdir_owned.as_deref().unwrap_or(bin_dir);
-    #[cfg(not(windows))]
-    let target_for_mkdir = bin_dir;
-    if let Err(e) = std::fs::create_dir_all(target_for_mkdir) {
-        let tolerated = e.kind() == std::io::ErrorKind::AlreadyExists && target_for_mkdir.is_dir();
-        if !tolerated {
-            return Err(e)
-                .into_diagnostic()
-                .wrap_err_with(|| format!("failed to create bin directory {}", bin_dir.display()));
-        }
-    }
-    aube_linker::create_bin_shim(target_for_mkdir, name, target, shim_opts)
-        .into_diagnostic()
-        .wrap_err_with(|| {
-            format!(
-                "failed to link bin `{name}` at {} -> {}",
-                bin_dir.join(name).display(),
-                target.display()
-            )
-        })?;
-    Ok(())
 }
